Teach plan  branch

+++ src/main/java/com/peter/kist/repository/TeacherPlanRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface TeacherPlanRepository extends JpaRepository<TeacherPlan, Long> {
+public interface TeacherPlanRepository extends JpaRepository<TeacherPlan, Integer> {
     @Modifying
     @Transactional
     void deleteTeacherPlanById(Integer id);
